% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albedo_list.R
\docType{data}
\name{albedo_list}
\alias{albedo_list}
\title{Albedo solution kernel values}
\format{A list with 26 elements, each element is a data frame, organized by site. See the file process-gsvd-fluxnet.R in the data-raw folder}
\usage{
data(albedo_list)
}
\description{
\itemize{
  \item time. Day of the year in 2017
  \item band. Reflectance band (1-7)
  \item value. Albedo measurement
  \item type. Black sky or white sky albedo. Currently only contains white-sky albedo - for black-sky albedo we need solar noon
}
}
\keyword{datasets}
