% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_uncertainty.R
\docType{data}
\name{band_uncertainty}
\alias{band_uncertainty}
\title{Uncertainty measurements on reflectance bands}
\format{A data frame with 7 rows and two columns}
\source{
Tristan
}
\usage{
data(band_uncertainty)
}
\description{
\itemize{
  \item band. Reflectance band
  \item sigma. Uncertainty, indexed by band
}
}
\keyword{datasets}
