% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd_solution.R
\name{gsvd_solution}
\alias{gsvd_solution}
\title{Compute the solution for the kernel weights for a GSVD decomposition}
\usage{
gsvd_solution(gsvdResult, lambda, data)
}
\arguments{
\item{gsvdResult}{decomposition}

\item{lambda}{smoothness parameter - is a list of values}

\item{data}{right hand side of the equation - is a list of values with the band, value, and measurement}
}
\value{
Solution to Krylov subspace solution
}
\description{
\code{gsvd_solution} Computes solution for a GSVD matrix decomposition
}
\examples{

# To be filled in later
}
