% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_matrix.R
\name{kernel_matrix}
\alias{kernel_matrix}
\title{Compute the matrix for BRDF inversion}
\usage{
kernel_matrix(site_name)
}
\arguments{
\item{site_name}{Name of the site}
}
\value{
A list that contains the K, B, and rho matrix, and time points
}
\description{
\code{kernel_matrix} Computes the B, K, and rho matrices for a given site in preparation to do the GSVD
}
\examples{

kernel_data <- kernel_matrix('AU-Lox')
}
