% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_list.R
\docType{data}
\name{lambda_list}
\alias{lambda_list}
\title{Max smoothing lambda for 26 MODIS pixels in 2017.}
\format{A list with 26 elements, organized by site. Each element is a data frame contaitning a max lambda value, organized by band and if it converged. See the file process-gsvd-fluxnet.R in the data-raw folder}
\usage{
data(lambda_list)
}
\description{
\itemize{
  \item band. Which band used
  \item lambda. Max lambda value
  \item converged. Were we able to find a value of lambda within the RMSE?
}
}
\keyword{datasets}
