% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_solution.R
\name{ls_solution}
\alias{ls_solution}
\title{Compute the solution for the kernel weights for a least squares decomposition}
\usage{
ls_solution(site_name)
}
\arguments{
\item{site_name}{Name of the site}
}
\value{
Least squares weighted solution
}
\description{
\code{ls_solution} Computes solution for a GSVD matrix decomposition
}
\examples{

# To be filled in later
}
