% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_norm.R
\name{max_norm}
\alias{max_norm}
\title{Compute the relative proximity of a solution in relation to an uncertainty.}
\usage{
max_norm(norm_result, uncertainty)
}
\arguments{
\item{norm_result}{Result of doing the gsvd_norm function}

\item{uncertainty}{Target level of the norm we want to acheive}
}
\value{
Data frame that contains the updated lambda min and max values
}
\description{
\code{max_norm} Determines where a norm (residual or solution) is within a certain tolerance, iteratively narrowing down a list of lambda values.
}
\examples{

# To be filled in later
}
