% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcd43A3.R
\docType{data}
\name{mcd43A3}
\alias{mcd43A3}
\title{White and black sky albedo for 26 MODIS pixels.}
\format{A data frame with 69080 rows and 6 variables}
\source{
Schaaf, C., Wang, Z. (2015). MCD43A3 MODIS/Terra+Aqua BRDF/Albedo Daily L3 Global - 500m V006. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MCD43A3.006. Accessed September 19, 2018.
}
\usage{
data(mcd43A3)
}
\description{
\itemize{
  \item site. Fluxnet site abbreviation
  \item time. Day of the year in 2017
  \item method. Data product name from MODIS
  \item band. Reflectance band (1-7)
  \item albedo. Albedo measurement
  \item type. Black sky or white sky albedo
}
}
\keyword{datasets}
