% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modisBRDF.R
\docType{data}
\name{modisBRDF}
\alias{modisBRDF}
\title{BRDF kernel weights for 26 MODIS pixels.}
\format{A data frame with 103620 rows and 5 variables}
\source{
Original Data:

  Schaaf, C., Wang, Z. (2015). MCD43A1 MODIS/Terra+Aqua BRDF/Albedo Model Parameters Daily L3 Global - 500m V006. NASA EOSDIS Land Processes DAAC. doi: 10.5067/MODIS/MCD43A1.006. Accessed December 20, 2018.
}
\usage{
data(modisBRDF)
}
\description{
\itemize{
  \item site. Fluxnet site abbreviation
  \item time. Day of the year in 2017
  \item value. Calculated value of kernel weight
  \item kernel. Code for kernel 0 = Isometric, 1 = Ross Thick kernel, 2 = Li Sparse kernel
  \item band. Reflectance bandNUM, where NUM is (1-7)
}
}
\keyword{datasets}
