% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_list.R
\docType{data}
\name{solution_list}
\alias{solution_list}
\title{Solution kernel weights for max smoothing lambda for 26 MODIS pixels in 2017.}
\format{A list with 26 elements, each element is a data frame, organized by site. See the file process-gsvd-fluxnet.R in the data-raw folder}
\usage{
data(solution_list)
}
\description{
\itemize{
  \item band. Reflectance bandNUM, where NUM is (1-7)
  \item time. Day of the year in 2017
  \item value. Calculated value of kernel weight
  \item kernel. Code for kernel 0 = Isometric, 1 = Ross Thick kernel, 2 = Li Sparse kernel
}
}
\keyword{datasets}
