% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albedo_compute.R
\name{albedo_compute}
\alias{albedo_compute}
\title{Compute the white sky albedo for an estimated kernel weight}
\source{
\url{https://www.umb.edu/spectralmass/terra_aqua_modis/v006/mcd43a1_brdif_albedo_model_parameters_product}
}
\usage{
albedo_compute(solution_data)
}
\arguments{
\item{solution}{A data frame of solution values, generated by gsvd_solution.  This has columns band, value, time, kernel}
}
\value{
Data frame that contains the time, band, value, and type (currently white-sky) albedo - for black-sky albedo we need solar noon
}
\description{
\code{albedo_compute} Computes the albedo values for a given kernel BRDF weights.
}
\examples{

# To be filled in later
}
